+++ Samples/shell/src/RendererExtensions.cpp
@@ -35,6 +35,10 @@
 
 	#include <RmlUi_Renderer_VK.h>
 
+#elif defined RMLUI_RENDERER_SDL_GPU
+
+	#include <RmlUi_Renderer_SDL_GPU.h>
+
 #endif
 
 #if defined RMLUI_RENDERER_GL2 || defined RMLUI_RENDERER_GL3
@@ -107,6 +111,27 @@ RendererExtensions::Image RendererExtensions::CaptureScreen()
 	if (!render_interface->CaptureScreen(image.width, image.height, image.num_components, image.data))
 	{
 		Rml::Log::Message(Rml::Log::LT_ERROR, "The Vulkan renderer was unable to capture screen.");
+		return {};
+	}
+
+	return image;
+}
+
+#elif defined RMLUI_RENDERER_SDL_GPU
+
+RendererExtensions::Image RendererExtensions::CaptureScreen()
+{
+	auto* render_interface = rmlui_static_cast<RenderInterface_SDL_GPU*>(Rml::GetRenderInterface());
+	if (!render_interface)
+	{
+		Rml::Log::Message(Rml::Log::LT_ERROR, "Could not capture screen. RmlUi render interface not set, or unexpected type.");
+		return {};
+	}
+
+	RendererExtensions::Image image;
+	if (!render_interface->CaptureScreen(image.width, image.height, image.num_components, image.data))
+	{
+		Rml::Log::Message(Rml::Log::LT_ERROR, "The SDL GPU renderer was unable to capture screen.");
 		return {};
 	}